+++ src/schema/distribuicaoDFe.js
@@ -1,19 +1,28 @@
 'use strict'
 
 exports.schema = (options) => {
+  const distDFeInt = {
+    tpAmb: options.tpAmb,
+    cUFAutor: options.cUFAutor,
+  }
+
+  if (options.cnpj) {
+    distDFeInt['CNPJ'] = options.cnpj
+  } else {
+    distDFeInt['CPF'] = options.cpf
+  }
+
+  distDFeInt[options.pesquisa.grupo] = {
+    [options.pesquisa.consulta]: options.pesquisa.valor,
+  }
+
+  distDFeInt['@_xmlns'] = 'http://www.portalfiscal.inf.br/nfe'
+  distDFeInt['@_versao'] = '1.01'
+
   return {
     nfeDistDFeInteresse: {
       nfeDadosMsg: {
-        distDFeInt: {
-          tpAmb: options.tpAmb,
-          cUFAutor: options.cUFAutor,
-          CNPJ: options.cnpj,
-          [options.pesquisa.grupo]: {
-            [options.pesquisa.consulta]: options.pesquisa.valor,
-          },
-          '@_xmlns': 'http://www.portalfiscal.inf.br/nfe',
-          '@_versao': '1.01',
-        },
+        distDFeInt: distDFeInt,
       },
       '@_xmlns': 'http://www.portalfiscal.inf.br/nfe/wsdl/NFeDistribuicaoDFe',
     },